greatly simplified the minimap component
it was relaying a lot of signals from other components.
now the only thing it does is create the minimap marker, the movement of the marker goes straight from the movement component, and the removal of the marker upon death goes straight from the health component.
also simplified the minimap code, now the camera directly controls the minimap camera too
183 hours